Update patch set 3
Patch Set 3:

Maybe prototype of vp8_dequant_dc_idct_add in dequantize.h could be changed to remove duplicate 'dst' and 'stride' parameters but that could be done in separate commit. This would affect at least c, mmx and armv6 functions.

It would also be possible to check, for example, in vp8_dequant_dc_idct_add_y_block_v6 function if eobs are zero for 4 blocks and skip one iteration of the loop if true.

Patch-set: 3