Update patch set 3
Patch Set 3:

I think you should change the function names to correspond with what they do now.  Currently it's filter block edge, but when you do 3x it's not the same.  It's a minor nit, but I think that the contents of the function have changed enough.  I also realize that this is an intern release and what I'm asking would require a lot of extra code because of the RTCD that would be removed in the future.

You may have missed an obscure configuration, 64 bit with mmx only.  It may not crash but will produce the wrong output.

Patch-set: 3
Label: Code-Review=0
Label: Verified=0