Update patch set 2
Patch Set 2: Looks good to me, approved

I think this is a better approach for this implementation, thanks.

As Jim mentioned, this implementation could be made a lot faster. If you're not going to work on that now, I think it'd be good if you could either expand the commit message or put a nearby comment describing what the "rare situations" are where this happens, to aid whomever is going to optimize this. For example, if this only happens during trellis, it might be easy to just return the updated eob in optimize_b and check that before optimizing the y2 block, avoiding a slow is_everything_zero() check. I'd like to be able to understand the context in which this can happen and hopefully fix it there, rather than losing that context and having to unconditionally check after the fact.

If we don't have any control over when this can happen and it'll always have to be done as a check on the data after the fact, I guess no further change needed, we can trivially write a faster zero check in a later commit. Just want to understand where the biggest opportunity is. Did you try this by checking the eobs before you checked the  data itself?

Patch-set: 2
Label: Code-Review=+2
Label: Verified=0