Update patch set 1
Patch Set 1:

No, it's to check if you're on 32 bit, and use -f win32, and on 64 bit, use -f x64 (instead of win64). We just haven't gotten around to writing the check and committing it. In the past, it hasn't required manual intervention:
 http://code.google.com/p/webm/issues/detail?id=232

If you'd like to write the if:else, we will take the patch

Patch-set: 1