Update patch set 6
Patch Set 6: (1 inline comment)

Installing the new ndk and the sort no longer works because the version number is appended to a string and the version has now gone over 10.  Looks a little ugly, open to other suggestions.

win doesn't build, a lot of it looks like path issues.  would prefer to fix in a later patch.

Patch-set: 6